Fix/Result 뒤로가기 오류핵결 및 Refactoring
